rejig component managers in EUDEMO
